+++ CHANGELOG.md
@@ -55,7 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.4.0...HEAD
 [0.4.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.1.0...0.2.0