+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Stripe promotion code function.
 - Stripe webhooks functions.
+- Data expansion feature to Stripe `checkoutSession` function.
 
 ### Changed
 - Replace checkout line items deprecated data parameters with `price_data`.