+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stripe promotion code function.
 - Stripe webhooks functions.
 - Data expansion feature to Stripe `checkoutSession` function.
+- Stripe checkout session creation `mode` to payload data.
 
 ### Changed
 - Replace checkout line items deprecated data parameters with `price_data`.