+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Replace checkout line items deprecated data parameters with `price_data`.
+- Bump `stripe` version number to `8.107.0`.
 
 ## [0.4.0] - 2020-10-06
 ### Added