+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2020-11-04
 ### Added
 - `.env` file to gitignore.
 - An improved router module.
@@ -78,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/my-jam-store/stripe-checkout-service/compare/0.2.0...0.3.0